lavet en del ændringer

+++ src/main/java/org/example/ailoldraftingcheck/entity/ApiUsageRepository.java
@@ -2,5 +2,5 @@ package org.example.ailoldraftingcheck.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApiUsageRepository extends JpaRepository<ApiUsage, Long> {
+public interface ApiUsageRepository extends JpaRepository<ApiUsage, Integer> {
 }
